Programming
Backpack was acting kind of weird so I decided to just replace it with a plank spawner. Basically you can add and remove planks to a spawner over your left shoulder now.
Also fixed the distance grab component so that you can't grab things out of your hands, or things that are too close to or far away from the player.